+++ src/api/admin.js
@@ -1,0 +1,1 @@
+import request from "./request";
