--- src/api/banner.js
@@ -1,5 +1,0 @@
-import request from "./request";
-
-export const loadBanner = (bannerCode)=> {
-  return request("display/banners/"+bannerCode, "get", null, null, {});
-};
