--- src/api/category.js
@@ -1,5 +1,0 @@
-import request from "./request";
-
-export const getCategoryList = (categoryNo, orderBy)=> {
-  return request("categories", "get", null, null, {});
-};
