+++ src/api/claim.js
@@ -1,0 +1,1 @@
+import request from "./request";
