--- src/api/events.js
@@ -1,5 +1,0 @@
-import request from './request';
-
-export const getEventByEventNo = (eventNo) => {
-  return request(`display/events/${eventNo}`, 'get', null, null, {});
-};
