+++ src/api/manage.js
@@ -1,5 +1,6 @@
 import request from "./request";
 
+// board
 export const getBoardList = (boardNo)=> {
   return request("boards/"+boardNo+"/articles?pageSize=10", "get", null, null, {});
 };
