+++ src/pages/error/error404.js
@@ -4,8 +4,7 @@
  import SEOHelmet from '../../components/SEOHelmet';
  
  //api
- import { sampleApi } from "../../api/sample";
- 
+
  //css
  import "../../assets/scss/contents.scss"
