+++ src/pages/event/liveon.js
@@ -4,7 +4,6 @@ import { React } from 'react';
 import SEOHelmet from '../../components/SEOHelmet';
 
 //api
-import { sampleApi } from "../../api/sample";
 
 //css
 import "../../assets/scss/contents.scss"
